Remove kthread and unblock user process from ISR
The kthread was introduced because you can't block or unblock on
mutexes inside a hard ISR, like a timer callback. So the timer
callback just *completed* and the kernel thread unblocked and did
the job of checking whether a user was blocked and *completing*
to unblock it.

Since the double inter thread communication (ISR to kthread,
kthread to blocked process) introduced some latency and it was
inefficient, to protect the blocked_user flag variable I used a
spinlock instead of a mutex, so that we can check it inside the
hard ISR.