Support multiple threads sharing file descriptor
Multiple threads share the same file descriptor, so if we want
them to use it without interfering with each other, we must track
every thread.

When a thread opens the device, a configuration list is
initialized. Every
process that wants to stop and periodically be woken, must set a
timeout using and ioctl(fd, TIMEOUT_SET, tms) call. This call will
add a new timer configuration (struct periodic_conf) to the list
or will update an existing configuration.

Every timer configuration stores the struct pid pointer of the
thread the configuration belongs to. We used the struct pid *
instead of the struct task_struct * as recommended in
<linux/pid.h:19-44>. The struct pid pointer is used to find the
correct configuration out of all the list across different calls
to the module functions performed by the same thread.

The config list is kept sorted so that configurations with shorter
timeouts are accessed more quickly, even though we did not exploit
this feature.